+++ application/controllers/HomeController.php
@@ -156,6 +156,18 @@ class HomeController extends MyAppController
                     }
                     $collectionTemplates[$collection['collection_id']]['html'] = $homePageProdLayout3;
                     break;
+                case Collections::TYPE_PRODUCT_LAYOUT4:
+                    $homePageProdLayout4 = CacheHelper::get('homePageProdLayout4' . $collection['collection_id'] . $cacheKey, CONF_HOME_PAGE_CACHE_TIME, '.txt');
+                    if (!$homePageProdLayout4) {
+                        $tpl = new FatTemplate('', '');
+                        $tpl->set('siteLangId', $this->siteLangId);
+                        $tpl->set('collection', $collection);
+                        $tpl->set('displayProductNotAvailableLable', $displayProductNotAvailableLable);
+                        $homePageProdLayout4 = $tpl->render(false, false, '_partial/collection/product-layout-4.php', true, true);
+                        CacheHelper::create('homePageProdLayout4' . $collection['collection_id'] . $cacheKey, $homePageProdLayout4, CacheHelper::TYPE_COLLECTIONS);
+                    }
+                    $collectionTemplates[$collection['collection_id']]['html'] = $homePageProdLayout4;
+                    break;
                 case Collections::TYPE_CATEGORY_LAYOUT1:
                     $homePageCatLayout1 = CacheHelper::get('homePageCatLayout1' . $collection['collection_id'] . $cacheKey, CONF_HOME_PAGE_CACHE_TIME, '.txt');
                     if (!$homePageCatLayout1) {
@@ -191,6 +203,18 @@ class HomeController extends MyAppController
                         CacheHelper::create('homePageCatLayout3' . $collection['collection_id'] . $cacheKey, $homePageCatLayout3, CacheHelper::TYPE_COLLECTIONS);
                     }
                     $collectionTemplates[$collection['collection_id']]['html'] = $homePageCatLayout3;
+                    break;
+                case Collections::TYPE_CATEGORY_LAYOUT4:
+                    $homePageCatLayout4 = CacheHelper::get('homePageCatLayout4' . $collection['collection_id'] . $cacheKey, CONF_HOME_PAGE_CACHE_TIME, '.txt');
+                    if (!$homePageCatLayout4) {
+                        $tpl = new FatTemplate('', '');
+                        $tpl->set('siteLangId', $this->siteLangId);
+                        $tpl->set('collection', $collection);
+                        $tpl->set('displayProductNotAvailableLable', $displayProductNotAvailableLable);
+                        $homePageCatLayout4 = $tpl->render(false, false, '_partial/collection/category-layout-4.php', true, true);
+                        CacheHelper::create('homePageCatLayout4' . $collection['collection_id'] . $cacheKey, $homePageCatLayout4, CacheHelper::TYPE_COLLECTIONS);
+                    }
+                    $collectionTemplates[$collection['collection_id']]['html'] = $homePageCatLayout4;
                     break;
                 case Collections::TYPE_SHOP_LAYOUT1:
                     $homePageShopLayout1 = CacheHelper::get('homePageShopLayout1' . $collection['collection_id'] . $cacheKey, CONF_HOME_PAGE_CACHE_TIME, '.txt');
@@ -779,7 +803,7 @@ class HomeController extends MyAppController
                     /* ] */
                     $collections[$ind] = $collection;
                     $counter = 0;
-                    if ($collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT2 || $collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT3) {
+                    if (in_array($collection['collection_layout_type'], [Collections::TYPE_CATEGORY_LAYOUT2, Collections::TYPE_CATEGORY_LAYOUT3])) {
                         while ($catData = $db->fetch($rs)) {
                             if (true === MOBILE_APP_API_CALL) {
                                 $imgUpdatedOn = ProductCategory::getAttributesById($catData['prodcat_id'], 'prodcat_updated_on');
@@ -807,7 +831,7 @@ class HomeController extends MyAppController
                                 $collections[$ind]['categories'][$catData['prodcat_id']]['subCategories'] = $subCategories;
 
                                 $product = [];
-                                if ($collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT3) {
+                                if ($collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT3 || $collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT4) {
                                     $allCats = [$catData['prodcat_id']] + array_keys($subCategories);
                                     $prodObj = clone $productSrchObj;
                                     $prodObj->addCondition('prodcat_id', 'IN', $allCats);
@@ -821,7 +845,7 @@ class HomeController extends MyAppController
                             /* ] */
                             $counter++;
                         }
-                    } else if ($collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT1) {
+                    } else if (in_array($collection['collection_layout_type'], [Collections::TYPE_CATEGORY_LAYOUT1, Collections::TYPE_CATEGORY_LAYOUT4])) {
                         while ($catData = $db->fetch($rs)) {
                             /* fetch Product data[ */
                             $productShopSrchTempObj = clone $productSrchObj;
@@ -829,7 +853,8 @@ class HomeController extends MyAppController
                             $productShopSrchTempObj->addCondition('prodcat_id', '=', $catData['prodcat_id']);
                             //$productShopSrchTempObj->addOrder('in_stock', 'DESC');
                             $productShopSrchTempObj->addGroupBy('selprod_product_id');
-                            $productShopSrchTempObj->setPageSize(Collections::LIMIT_CATEGORY_LAYOUT1_PRODUCT);
+                            $limit = ($collection['collection_layout_type'] == Collections::TYPE_CATEGORY_LAYOUT1) ? Collections::LIMIT_CATEGORY_LAYOUT1_PRODUCT : Collections::LIMIT_CATEGORY_LAYOUT4;
+                            $productShopSrchTempObj->setPageSize($limit);
 
                             $recordCount = $this->getRecordsCount(clone $productShopSrchTempObj, true);
                             if (empty($recordCount)) {