+++ application/views/cart/ship-listing.php
@@ -139,9 +139,9 @@ if (UserAuthentication::isUserLogged() && (!User::isBuyer())) {
                                 </div>
                                 <a class="title" href="<?php echo $productUrl; ?>"><?php echo $productTitle; ?></a>
                                 <div class="products-price">
-                                    <span class="products-price-new"><?php echo trim(CommonHelper::displayMoneyFormat($product['theprice'])); ?></span>
+                                    <span class="products-price-new"><?php echo trim(CommonHelper::displayMoneyFormat($product['theprice'], true, false, true, false, false, true)); ?></span>
                                     <?php if ($product['special_price_found'] && $product['selprod_price'] > $product['theprice']) { ?>
-                                        <del class="products-price-old"><?php echo trim(CommonHelper::displayMoneyFormat($product['selprod_price'])); ?></del>
+                                        <del class="products-price-old"><?php echo trim(CommonHelper::displayMoneyFormat($product['selprod_price'], true, false, true, false, false, true)); ?></del>
                                         <div class="products-price-off"><?php echo trim(CommonHelper::showProductDiscountedText($product, $siteLangId)); ?></div>
                                     <?php } ?>
                                 </div>
@@ -269,7 +269,7 @@ if (UserAuthentication::isUserLogged() && (!User::isBuyer())) {
                                     </div>
                                     <a class="title" href="<?php echo $productUrl; ?>"><?php echo $productTitle; ?></a>
                                     <div class="products-price">
-                                        <?php echo CommonHelper::displayMoneyFormat($product['theprice']); ?></div>
+                                        <?php echo CommonHelper::displayMoneyFormat($product['theprice'], true, false, true, false, false, true); ?></div>
                                     <div class="options">
                                         <?php
                                         if (isset($product['options']) && count($product['options'])) {