+++ application/views/checkout/shipping-summary-group.php
@@ -69,9 +69,9 @@
                             <a class="title" href="<?php echo $productUrl; ?>"><?php echo ($product['selprod_title']) ? $product['selprod_title'] : $product['product_name']; ?></a>
 
                             <div class="products-price">
-                                <span class="products-price-new"><?php echo trim(CommonHelper::displayMoneyFormat($product['actualPrice'])); ?></span>
+                                <span class="products-price-new"><?php echo trim(CommonHelper::displayMoneyFormat($product['actualPrice'], true, false, true, false, false, true)); ?></span>
                                 <?php if ($product['selprod_price'] > $product['actualPrice']) { ?>
-                                    <del class="products-price-old"><?php echo trim(CommonHelper::displayMoneyFormat($product['selprod_price'])); ?></del>
+                                    <del class="products-price-old"><?php echo trim(CommonHelper::displayMoneyFormat($product['selprod_price'], true, false, true, false, false, true)); ?></del>
                                     <div class="products-price-off"><?php echo trim(CommonHelper::showProductDiscountedText($product, $siteLangId, 'actualPrice')); ?></div>
                                 <?php } ?>
                             </div>