+++ application/views/checkout/shipping-summary-pickup.php
@@ -167,9 +167,9 @@
                                                     <?php echo ($product['selprod_title']) ? $product['selprod_title'] : $product['product_name']; ?>
                                                 </a>
                                                 <div class="products-price">
-                                                    <span class="products-price-new"><?php echo trim(CommonHelper::displayMoneyFormat($product['theprice'])); ?></span>
+                                                    <span class="products-price-new"><?php echo trim(CommonHelper::displayMoneyFormat($product['theprice'], true, false, true, false, false, true)); ?></span>
                                                     <?php if ($product['special_price_found'] && $product['selprod_price'] > $product['theprice']) { ?>
-                                                        <del class="products-price-old"><?php echo trim(CommonHelper::displayMoneyFormat($product['selprod_price'])); ?></del>
+                                                        <del class="products-price-old"><?php echo trim(CommonHelper::displayMoneyFormat($product['selprod_price'], true, false, true, false, false, true)); ?></del>
                                                         <div class="products-price-off"><?php echo trim(CommonHelper::showProductDiscountedText($product, $siteLangId)); ?></div>
                                                     <?php } ?>
                                                 </div>
@@ -260,7 +260,7 @@
                                             <div class="product-profile-data">
                                                 <a class="title" href="<?php echo $productUrl; ?>"><?php echo ($product['selprod_title']) ? $product['selprod_title'] : $product['product_name']; ?></a>
                                                 <div class="products-price">
-                                                    <?php echo CommonHelper::displayMoneyFormat($product['theprice'] * $product['quantity']); ?>
+                                                    <?php echo CommonHelper::displayMoneyFormat($product['theprice'] * $product['quantity'], true, false, true, false, false, true); ?>
                                                     <?php if ($product['selprod_price'] > $product['theprice']) { ?>
                                                         <span class="products-price-off">
                                                             <?php echo CommonHelper::showProductDiscountedText($product, $siteLangId); ?></span>